refactor Mission 3 and remove HeadEstimator
