+++ frontend/src/pages/InGame/Mission3/Mission3.js
@@ -4,8 +4,8 @@ import {
   GameContext,
   OpenViduContext,
 } from '../../../contexts';
+import * as pose from '@mediapipe/pose';
 import { MissionStarting, MissionEnding } from '../components';
-import { estimateHead } from '../MissionEstimators/HeadEstimator';
 import arrow from '../../../assets/arrows/arrow.svg';
 import { RoundSoundEffect, MissionSoundEffects } from '../Sound';
 
@@ -25,6 +25,11 @@ const round2 = [
   { id: 7, direction: 'right', active: false },
 ];
 
+let currentStatus; // 현재 고개를 돌린 방향
+let isCentered = false;
+let isDirectionCorrect = false; // 화살표 별 측정 결과
+let isMissionFinished = false;
+
 const Mission3 = () => {
   const { poseModel, setIsPoseLoaded, setIsPoseInitialized } =
     useContext(MediaPipeContext);
@@ -38,7 +43,6 @@ const Mission3 = () => {
     setGameScore,
   } = useContext(GameContext);
   const { myVideoRef } = useContext(OpenViduContext);
-  const canvasRef = useRef(null);
 
   // 화살표 세팅
   const [arrowRound, setArrowRound] = useState({
@@ -47,13 +51,90 @@ const Mission3 = () => {
   });
   const [currentRoundIdx, setCurrentRoundIdx] = useState(0);
   const [currentArrowIdx, setCurrentArrowIdx] = useState(0);
-  const [isMissionFinished, setIsMissionFinished] = useState(false);
   const score = useRef(0);
+
+  const NeckGame = (poseLandmarks, direction) => {
+    if (!poseLandmarks) return;
+
+    // 각 포인트 별 위치 기록
+    const nose = poseLandmarks[pose.POSE_LANDMARKS.NOSE];
+    const leftEar = poseLandmarks[pose.POSE_LANDMARKS.LEFT_EAR];
+    const rightEar = poseLandmarks[pose.POSE_LANDMARKS.RIGHT_EYE];
+
+    // ========= ⭐️⭐️⭐️ ==========
+    // fps 안 나오면 해당 부분 단순화 예정
+    const leftShoulder = poseLandmarks[pose.POSE_LANDMARKS.LEFT_SHOULDER];
+    const rightShoulder = poseLandmarks[pose.POSE_LANDMARKS.RIGHT_SHOULDER];
+    const leftMouth = poseLandmarks[pose.POSE_LANDMARKS.LEFT_RIGHT];
+    const rightMouth = poseLandmarks[pose.POSE_LANDMARKS.RIGHT_LEFT];
+    const centerSholderY = (leftShoulder.y + rightShoulder.y) / 2;
+    const centerMouthY = (leftMouth.y + rightMouth.y) / 2;
+
+    // 얼굴을 해당 방향으로 돌렸는지 확인
+    if (!isMissionFinished) {
+      isDirectionCorrect = false;
+      if (
+        nose.x < leftEar.x &&
+        nose.x > rightEar.x &&
+        centerSholderY - centerMouthY > (centerMouthY - nose.y) * 4.5 &&
+        isCentered
+      ) {
+        currentStatus = 'top';
+        isCentered = false;
+      } else if (
+        nose.x < leftEar.x &&
+        nose.x > rightEar.x &&
+        centerSholderY - centerMouthY < (centerMouthY - nose.y) * 1.9 &&
+        isCentered
+      ) {
+        currentStatus = 'bottom';
+        isCentered = false;
+      } else if (
+        nose.x > leftEar.x &&
+        nose.x > rightEar.x &&
+        Math.abs(leftEar.x - nose.x) >= Math.abs(rightEar.x - leftEar.x) &&
+        isCentered
+      ) {
+        currentStatus = 'right';
+        isCentered = false;
+      } else if (
+        nose.x < leftEar.x &&
+        nose.x < rightEar.x &&
+        // Math.abs(rightEar.x - nose.x) <= Math.abs(rightEar.x - leftEar.x) &&
+        isCentered
+      ) {
+        currentStatus = 'left';
+        isCentered = false;
+      } else if (
+        nose.x < leftEar.x &&
+        nose.x > rightEar.x &&
+        centerSholderY - centerMouthY < (centerMouthY - nose.y) * 4 &&
+        centerSholderY - centerMouthY > (centerMouthY - nose.y) * 2.2
+        // && nose.y > leftShoulder.y - rangeShoulderY &&
+        // nose.y < leftShoulder.y + rangeShoulderY
+      ) {
+        // 한 번 고개를 돌린 뒤에는 정면을 봐야 점수를 줌
+        currentStatus = 'front';
+        isCentered = true;
+      }
+
+      if (
+        // currentStatus === selectedDirection[0]
+        currentStatus === direction
+      ) {
+        // 현재 상태가 다음 방향과 일치하는지 확인
+        // selectedDirection.shift(); // 다음 방향으로 이동
+        isDirectionCorrect = true;
+      }
+
+      return isDirectionCorrect;
+    }
+  };
 
   useEffect(() => {
     const timer = setTimeout(() => {
       setGameScore(prev => prev + score.current);
-      setIsMissionFinished(true);
+      isMissionFinished = true;
     }, 17000);
     return () => clearTimeout(timer); // 컴포넌트 언마운트 시 타이머 클리어
   }, []);
@@ -92,29 +173,15 @@ const Mission3 = () => {
   }, [isMissionStarting, poseModel]);
 
   useEffect(() => {
-    if (isMissionFinished) {
-      return;
-    }
+    if (isMissionFinished) return;
     if (!poseModel.current || isMissionStarting) return;
 
     const direction = arrowRound[currentRoundIdx][currentArrowIdx].direction;
-    if (myMissionStatus) {
-      score.current = 25;
-    } else if (currentRoundIdx === 1) {
-      score.current = 12 + (currentArrowIdx + 1) * 3 - 3;
-    } else if (currentRoundIdx === 0) {
-      score.current = (currentArrowIdx + 1) * 3 - 3;
-    }
-    // console.log('score : ', score.current);
-    poseModel.current.onResults(results => {
-      const result = estimateHead({
-        results,
-        myVideoRef,
-        canvasRef,
-        direction,
-      });
 
-      if (result) {
+    poseModel.current.onResults(results => {
+      const isCorrect = NeckGame(results.poseLandmarks, direction);
+
+      if (isCorrect) {
         setArrowRound(prevState => {
           const newState = { ...prevState };
           newState[currentRoundIdx][currentArrowIdx].active = true;
@@ -126,11 +193,13 @@ const Mission3 = () => {
           setMyMissionStatus(true); // 성공
           score.current = 25;
         } else if (currentRoundIdx === 0 && currentArrowIdx === 3) {
+          score.current = 12;
           setCurrentArrowIdx(0); // 첫 번째 화살표로 초기화
           setTimeout(() => {
             setCurrentRoundIdx(currentRoundIdx + 1); // 다음 라운드로 넘어감
           }, 1000); // 1초 뒤에 실행되도록 설정
         } else {
+          score.current = currentRoundIdx * 12 + (currentArrowIdx + 1) * 3;
           setCurrentArrowIdx(currentArrowIdx + 1); // 다음 화살표로 이동
         }
       }
@@ -158,7 +227,6 @@ const Mission3 = () => {
       {isMissionEnding && <MissionSoundEffects />}
       {isMissionStarting || (
         <>
-          <Canvas ref={canvasRef} />
           <ArrowBox>
             {arrowRound[currentRoundIdx].map(({ id, direction, active }) => (
               <Arrows
@@ -177,13 +245,6 @@ const Mission3 = () => {
 };
 
 export default Mission3;
-
-const Canvas = styled.canvas`
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  border-radius: ${({ theme }) => theme.radius.medium};
-`;
 
 const ArrowBox = styled.div`
   z-index: 200;
